+++ pkg/api/client/basic/client_test.go
@@ -21,20 +21,18 @@ import (
 func Test_StepRunnerClient_Status_ListJobs(t *testing.T) {
 	rr1 := test.RunRequest(t, `run:
   - name: hello_world
-    step: ../../../../e2e_tests/steps/greeting
-    inputs: {}
+    script: echo "hello world"
 `, nil, nil)
 	rr1.Id = rr1.Id + "-1"
 
 	rr2 := test.RunRequest(t, `run:
   - name: blabla
-    step: ../../testdata/bash
-    inputs:
-        script: echo "bla bla bla"
+    script: echo "bla bla bla"
 `, nil, nil)
 	rr1.Id = rr1.Id + "-2"
 
 	server := server.New(t).Serve()
+	defer server.Stop()
 	srClient := New(server.NewConnection())
 
 	ctx := context.Background()
@@ -74,11 +72,10 @@ const (
 func Test_StepRunnerClient_FollowLogs_Success(t *testing.T) {
 	rr := test.RunRequest(t, `run:
   - name: lorem
-    step: ../../testdata/bash
-    inputs:
-        script: echo "`+lorem+`"`, nil, nil)
+    script: echo "`+lorem+`"`, nil, nil)
 
 	server := server.New(t).Serve()
+	defer server.Stop()
 	srClient := New(server.NewConnection())
 
 	ctx := context.Background()
@@ -90,6 +87,7 @@ func Test_StepRunnerClient_FollowLogs_Success(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, int64(len(runStepMsg)+len(lorem)+1), n)
 	assert.Equal(t, runStepMsg+lorem+"\n", buf.String())
+
 	assert.NoError(t, srClient.Close(ctx, rr.Id))
 }
 
@@ -100,11 +98,10 @@ func (t toWriter) Write(p []byte) (int, error) { return t(p) }
 func Test_StepRunnerClient_FollowLogs_Again(t *testing.T) {
 	rr := test.RunRequest(t, `run:
   - name: lorem
-    step: ../../testdata/bash
-    inputs:
-        script: echo "`+lorem+`"`, nil, nil)
+    script: echo "`+lorem+`"`, nil, nil)
 
 	server := server.New(t).Serve()
+	defer server.Stop()
 	srClient := New(server.NewConnection())
 
 	ctx := context.Background()
@@ -141,8 +138,9 @@ func Test_StepRunnerClient_WaitForReady(t *testing.T) {
 		conn := srvr.NewConnection()
 		require.Eventually(t, func() bool { return conn.GetState() == connectivity.Idle }, 2*time.Second, 100*time.Millisecond)
 
-		step := "run:\n  - name: hello_world\n    step: ../../../../e2e_tests/steps/greeting"
-		runRequest := test.RunRequest(t, step, nil, nil)
+		runRequest := test.RunRequest(t, `run:
+      - name: hello_world
+		script: echo "ghello wrold"`, nil, nil)
 
 		ctx, cancel := context.WithTimeout(context.Background(), time.Millisecond*50)
 		defer cancel()
@@ -158,10 +156,16 @@ func Test_StepRunnerClient_WaitForReady(t *testing.T) {
 		require.Eventually(t, func() bool { return conn.GetState() == connectivity.Idle }, 2*time.Second, 100*time.Millisecond)
 
 		srvr.Serve()
-		step := "run:\n  - name: hello_world\n    step: ../../../../e2e_tests/steps/greeting"
-		runRequest := test.RunRequest(t, step, nil, nil)
+		defer srvr.Stop()
+
+		runRequest := test.RunRequest(t, `run:
+      - name: hello_world
+        script: echo "hello world"`, nil, nil)
 
 		srClient := New(conn)
-		require.NoError(t, srClient.Run(context.Background(), runRequest))
+
+		ctx := context.Background()
+		require.NoError(t, srClient.Run(ctx, runRequest))
+		assert.NoError(t, srClient.Close(ctx, runRequest.Id))
 	})
 }